+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project uses [Semantic Versioning](https://semver.org).
 
 ## [Unreleased][unreleased]
+* Added the connection class and the basic response object
+* Added a filter for sensitive data to vcr
 * Added the base classes
 * Added support for dev environments, vcr and webmock as well configured rubocop and added a class to configure the gem itself
 * initial commit